# api/post/controller.go
package post

import (
	"sync-backend/api/common/analytics"
	modMW "sync-backend/api/moderator/middleware"
	"sync-backend/api/post/dto"
	"sync-backend/api/post/model"
	"sync-backend/arch/common"
	"sync-backend/arch/middleware"
	"sync-backend/arch/network"
	"sync-backend/utils"

	"github.com/gin-gonic/gin"
)

type postController struct {
	network.BaseController
	common.ContextPayload
	authenticatorProvider network.AuthenticationProvider
	uploadProvider        middleware.UploadProvider
	moderatorMiddleware   modMW.ModeratorMiddleware
	logger                utils.AppLogger
	postService           PostService
	postAnalytics         analytics.PostAnalytics
	communityAnalytics    analytics.CommunityAnalytics
}

func NewPostController(authenticatorProvider network.AuthenticationProvider, uploadProvider middleware.UploadProvider, postService PostService, postAnalytics analytics.PostAnalytics, communityAnalytics analytics.CommunityAnalytics, moderatorMiddleware modMW.ModeratorMiddleware) *postController {
	return &postController{
		BaseController:        network.NewBaseController("/post", authenticatorProvider),
		ContextPayload:        common.NewContextPayload(),
		logger:                utils.NewServiceLogger("PostController"),
		authenticatorProvider: authenticatorProvider,
		uploadProvider:        uploadProvider,
		moderatorMiddleware:   moderatorMiddleware,
		postService:           postService,
		postAnalytics:         postAnalytics,
		communityAnalytics:    communityAnalytics,
	}
}

func (c *postController) MountRoutes(group *gin.RouterGroup) {
	c.logger.Info("Mounting post routes")
	group.Use(c.authenticatorProvider.Middleware())
	group.GET("/get/:postId", c.GetPost)
	group.POST("/create", c.uploadProvider.Middleware("media"), c.CreatePost)
	group.PUT("/:postId", c.EditPost)
	group.DELETE("/:postId", c.DeletePost)

	group.POST("/like/:postId", c.LikePost)
	group.POST("/dislike/:postId", c.DislikePost)
	group.POST("/save/:postId", c.SavePost)
	group.POST("/share/:postId", c.SharePost)

	// User post routes
	group.GET("/get/user", c.UserPosts)
	group.GET("/feed", c.UserFeedPosts)
	group.GET("/trending", c.TrendingPosts)
	group.GET("/popular", c.PopularPosts)
	group.GET("/saved", c.UserSavedPosts)

	// Community post routes
	group.GET("/get/community/:communityId", c.GetCommunityPosts)

	// Post moderation actions
	group.POST("/:postId/sticky", c.StickyPost)
	group.POST("/:postId/lock", c.LockPost)
	group.POST("/:postId/archive", c.ArchivePost)
	group.POST("/:postId/nsfw", c.MarkNSFW)
}

func (c *postController) CreatePost(ctx *gin.Context) {
	body, err := network.ReqForm(ctx, dto.NewCreatePostRequest())
	if err != nil {
		return
	}
	userId := c.MustGetUserId(ctx)
	files := c.uploadProvider.GetUploadedFiles(ctx, "media")
	var filesList []string
	if files != nil && len(files.Files) > 0 {
		for _, file := range files.Files {
			c.logger.Debug("File uploaded: %s", file.Path)
			filesList = append(filesList, file.Path)
		}
	}
	if len(filesList) > 10 {
		c.Send(ctx).BadRequestError(
			"Too many files uploaded",
			"The maximum number of files allowed is 10",
			nil,
		)
		return
	}
	post, err := c.postService.CreatePost(
		body.Title,
		body.Content,
		body.Tags,
		filesList,
		*userId,
		body.CommunityId,
		body.Type,
		body.IsNSFW,
		body.IsSpoiler,
	)
	if err != nil {
		c.Send(ctx).MixedError(err)
		return
	}

	c.Send(ctx).SuccessDataResponse("Post created successfully", dto.CreatePostResponse{PostId: post.PostId})
	c.logger.Debug("Post details: %+v", post)
	c.uploadProvider.DeleteUploadedFiles(ctx, "media")

	go c.communityAnalytics.RecordPostCreated(post.CommunityId, *userId)
}

func (c *postController) GetPost(ctx *gin.Context) {
	postId := ctx.Param("postId")
	if postId == "" {
		c.Send(ctx).BadRequestError(
			"Post ID is required",
			"Please provide a valid post ID in the request params.",
			nil,
		)
		return
	}

	userId := c.MustGetUserId(ctx)
	post, err := c.postService.GetPost(postId, *userId)
	if err != nil {
		c.Send(ctx).MixedError(err)
		return
	}
	c.Send(ctx).SuccessDataResponse("Post retrieved successfully", post)
	c.logger.Debug("Post details: %+v", post)

	go c.postService.RecordPostView(postId, *userId)
	go c.postAnalytics.RecordPostClick(postId, *userId)
}

func (c *postController) EditPost(ctx *gin.Context) {
	postId := ctx.Param("postId")
	if postId == "" {
		c.Send(ctx).BadRequestError(
			"Post ID is required",
			"Please provide a valid post ID in the request params.",
			nil,
		)
		return
	}
	body, err := network.ReqForm(ctx, dto.NewEditPostRequest())
	if err != nil {
		return
	}
	userId := c.MustGetUserId(ctx)
	title := body.Title
	content := body.Content
	isNSFW := body.IsNSFW
	isSpoiler := body.IsSpoiler
	_, err = c.postService.EditPost(
		*userId,
		postId,
		&title,
		&content,
		body.PostType,
		&isNSFW,
		&isSpoiler,
	)
	if err != nil {
		c.Send(ctx).MixedError(err)
		return
	}
	c.Send(ctx).SuccessMsgResponse("Post edited successfully")
}

func (c *postController) DeletePost(ctx *gin.Context) {
	postId := ctx.Param("postId")
	if postId == "" {
		c.Send(ctx).BadRequestError(
			"Post ID is required",
			"Please provide a valid post ID in the request params.",
			nil,
		)
		return
	}

	userId := c.MustGetUserId(ctx)
	err := c.postService.DeletePost(*userId, postId)
	if err != nil {
		c.Send(ctx).MixedError(err)
		return
	}
	c.Send(ctx).SuccessMsgResponse("Post deleted successfully")
}

func (c *postController) LikePost(ctx *gin.Context) {
	postId := ctx.Param("postId")
	if postId == "" {
		c.Send(ctx).BadRequestError(
			"Post ID is required",
			"Please provide a valid post ID in the request params.",
			nil,
		)
		return
	}

	userId := c.MustGetUserId(ctx)
	isLiked, synergy, err := c.postService.LikePost(*userId, postId)
	if err != nil {
		c.Send(ctx).MixedError(err)
		return
	}

	c.Send(ctx).SuccessDataResponse("Post liked successfully", dto.LikePostResponse{
		PostId:  postId,
		IsLiked: isLiked,
		Synergy: synergy,
	})

	go c.postAnalytics.RecordPostVote(postId, *userId, +1)
}

func (c *postController) DislikePost(ctx *gin.Context) {
	postId := ctx.Param("postId")
	if postId == "" {
		c.Send(ctx).BadRequestError(
			"Post ID is required",
			"Please provide a valid post ID in the request params.",
			nil,
		)
		return
	}
	userId := c.MustGetUserId(ctx)
	isDisliked, synergy, err := c.postService.DislikePost(*userId, postId)
	if err != nil {
		c.Send(ctx).MixedError(err)
		return
	}
	c.Send(ctx).SuccessDataResponse("Post disliked successfully", dto.DislikePostResponse{
		PostId:     postId,
		IsDisliked: isDisliked,
		Synergy:    synergy,
	})

	go c.postAnalytics.RecordPostVote(postId, *userId, -1)
}

func (c *postController) SavePost(ctx *gin.Context) {
	postId := ctx.Param("postId")
	if postId == "" {
		c.Send(ctx).BadRequestError(
			"Post ID is required",
			"Please provide a valid post ID in the request params.",
			nil,
		)
		return
	}
	userId := c.MustGetUserId(ctx)
	err := c.postService.SavePost(*userId, postId)
	if err != nil {
		c.Send(ctx).MixedError(err)
		return
	}
	c.Send(ctx).SuccessMsgResponse("Post saved successfully")

	go c.postAnalytics.RecordPostSave(postId, *userId)
}

func (c *postController) SharePost(ctx *gin.Context) {
	postId := ctx.Param("postId")
	if postId == "" {
		c.Send(ctx).BadRequestError(
			"Post ID is required",
			"Please provide a valid post ID in the request params.",
			nil,
		)
		return
	}

	userId := c.MustGetUserId(ctx)
	err := c.postService.SharePost(*userId, postId)
	if err != nil {
		c.Send(ctx).MixedError(err)
		return
	}
	c.Send(ctx).SuccessMsgResponse("Post shared successfully")

	go c.postAnalytics.RecordPostShare(postId, *userId)
}

func (c *postController) UserPosts(ctx *gin.Context) {
	userId := c.MustGetUserId(ctx)
	body, err := network.ReqQuery(ctx, dto.NewGetUserPostRequest())
	if err != nil {
		return
	}
	posts, numberPosts, err := c.postService.GetPostsByUserId(*userId, body.Page, body.Limit)
	if err != nil {
		c.Send(ctx).MixedError(err)
		return
	}

	postsValue := make([]model.Post, len(posts))
	for i, post := range posts {
		if post != nil {
			postsValue[i] = *post
		}
	}
	c.Send(ctx).SuccessDataResponse("User posts retrieved successfully", dto.NewGetUserPostResponse(postsValue, body.Page, body.Limit, numberPosts))

	for _, post := range posts {
		go c.postAnalytics.RecordPostView(post.PostId, *userId)
	}
}

func (c *postController) UserFeedPosts(ctx *gin.Context) {
	userId := c.MustGetUserId(ctx)
	body, err := network.ReqQuery(ctx, dto.NewGetUserFeedPostRequest())
	if err != nil {
		return
	}
	posts, err := c.postService.GetUserFeedPosts(*userId, body.Page, body.Limit)
	if err != nil {
		c.Send(ctx).MixedError(err)
		return
	}

	postsValue := make([]model.FeedPost, len(posts))
	for i, post := range posts {
		if post != nil {
			postsValue[i] = *post
		}
	}
	c.Send(ctx).SuccessDataResponse("User feed posts retrieved successfully", dto.NewGetUserFeedPostResponse(postsValue, body.Page, body.Limit))

	for _, post := range posts {
		go c.postService.RecordPostView(post.ID, *userId)
		go c.postAnalytics.RecordPostView(post.ID, *userId)
	}
}

func (c *postController) TrendingPosts(ctx *gin.Context) {
	body, err := network.ReqQuery(ctx, dto.NewGetTrendingPostRequest())
	if err != nil {
		return
	}

	userId := c.MustGetUserId(ctx)
	posts, err := c.postService.GetTrendingPosts(*userId, body.Page, body.Limit)
	if err != nil {
		c.Send(ctx).MixedError(err)
		return
	}

	postsValue := make([]model.FeedPost, len(posts))
	for i, post := range posts {
		if post != nil {
			postsValue[i] = *post
		}
	}
	c.Send(ctx).SuccessDataResponse("Trending posts retrieved successfully", dto.NewGetTrendingPostResponse(postsValue, body.Page, body.Limit))

	for _, post := range posts {
		go c.postService.RecordPostView(post.ID, *userId)
		go c.postAnalytics.RecordPostView(post.ID, *c.MustGetUserId(ctx))
	}
}

func (c *postController) PopularPosts(ctx *gin.Context) {
	body, err := network.ReqQuery(ctx, dto.NewGetPopularPostRequest())
	if err != nil {
		return
	}

	userId := c.MustGetUserId(ctx)
	posts, err := c.postService.GetPopularPosts(*userId, body.Page, body.Limit)
	if err != nil {
		c.Send(ctx).MixedError(err)
		return
	}

	postsValue := make([]model.FeedPost, len(posts))
	for i, post := range posts {
		if post != nil {
			postsValue[i] = *post
		}
	}
	c.Send(ctx).SuccessDataResponse("Popular posts retrieved successfully", dto.NewGetPopularPostResponse(postsValue, body.Page, body.Limit))

	for _, post := range posts {
		go c.postService.RecordPostView(post.ID, *userId)
		go c.postAnalytics.RecordPostView(post.ID, *userId)
	}
}

func (c *postController) UserSavedPosts(ctx *gin.Context) {
	userId := c.MustGetUserId(ctx)
	body, err := network.ReqQuery(ctx, dto.NewGetUserSavedPostRequest())
	if err != nil {
		return
	}
	posts, err := c.postService.GetUserSavedPosts(*userId, body.Page, body.Limit)
	if err != nil {
		c.Send(ctx).MixedError(err)
		return
	}

	postsValue := make([]model.FeedPost, len(posts))
	for i, post := range posts {
		if post != nil {
			postsValue[i] = *post
		}
	}
	c.Send(ctx).SuccessDataResponse("User saved posts retrieved successfully", dto.NewGetUserSavedPostResponse(postsValue, body.Page, body.Limit))
}

func (c *postController) GetCommunityPosts(ctx *gin.Context) {
	communityId := ctx.Param("communityId")
	if communityId == "" {
		c.Send(ctx).BadRequestError(
			"Community ID is required",
			"Please provide a valid community ID in the request params.",
			nil,
		)
		return
	}
	body, err := network.ReqQuery(ctx, dto.NewGetCommunityPostRequest())
	if err != nil {
		return
	}
	posts, numberPosts, err := c.postService.GetPostsByCommunityId(communityId, body.Page, body.Limit)
	if err != nil {
		c.Send(ctx).MixedError(err)
		return
	}

	postsValue := make([]model.Post, len(posts))
	for i, post := range posts {
		if post != nil {
			postsValue[i] = *post
		}
	}
	c.Send(ctx).SuccessDataResponse("Community posts retrieved successfully", dto.NewGetCommunityPostResponse(postsValue, body.Page, body.Limit, numberPosts))

	userId := c.MustGetUserId(ctx)
	for _, post := range posts {
		go c.postService.RecordPostView(post.PostId, *userId)
		go c.postAnalytics.RecordPostView(post.PostId, *userId)
	}
}

func (c *postController) StickyPost(ctx *gin.Context) {
	postId := ctx.Param("postId")
	if postId == "" {
		c.Send(ctx).BadRequestError("Post ID is required", "Please provide a valid post ID in the request params.", nil)
		return
	}
	userId := c.MustGetUserId(ctx)
	newValue, err := c.postService.ToggleStickyPost(*userId, postId)
	if err != nil {
		c.Send(ctx).MixedError(err)
		return
	}
	c.Send(ctx).SuccessDataResponse("Post sticky toggled successfully", dto.NewTogglePostResponse(postId, "isStickied", newValue))
}

func (c *postController) LockPost(ctx *gin.Context) {
	postId := ctx.Param("postId")
	if postId == "" {
		c.Send(ctx).BadRequestError("Post ID is required", "Please provide a valid post ID in the request params.", nil)
		return
	}
	userId := c.MustGetUserId(ctx)
	newValue, err := c.postService.ToggleLockPost(*userId, postId)
	if err != nil {
		c.Send(ctx).MixedError(err)
		return
	}
	c.Send(ctx).SuccessDataResponse("Post lock toggled successfully", dto.NewTogglePostResponse(postId, "isLocked", newValue))
}

func (c *postController) ArchivePost(ctx *gin.Context) {
	postId := ctx.Param("postId")
	if postId == "" {
		c.Send(ctx).BadRequestError("Post ID is required", "Please provide a valid post ID in the request params.", nil)
		return
	}
	userId := c.MustGetUserId(ctx)
	newValue, err := c.postService.ToggleArchivePost(*userId, postId)
	if err != nil {
		c.Send(ctx).MixedError(err)
		return
	}
	c.Send(ctx).SuccessDataResponse("Post archive toggled successfully", dto.NewTogglePostResponse(postId, "isArchived", newValue))
}

func (c *postController) MarkNSFW(ctx *gin.Context) {
	postId := ctx.Param("postId")
	if postId == "" {
		c.Send(ctx).BadRequestError("Post ID is required", "Please provide a valid post ID in the request params.", nil)
		return
	}
	userId := c.MustGetUserId(ctx)
	newValue, err := c.postService.ToggleNSFWPost(*userId, postId)
	if err != nil {
		c.Send(ctx).MixedError(err)
		return
	}
	c.Send(ctx).SuccessDataResponse("Post NSFW toggled successfully", dto.NewTogglePostResponse(postId, "isNSFW", newValue))
}