# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Sync Backend API
  description: |
    Comprehensive API documentation for Sync - a social networking platform backend.

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Features
    - User authentication and management
    - Community creation and moderation
    - Post creation and interactions
    - Comment system with threading
    - File uploads for media content

  version: 1.0.0
  contact:
    name: API Support
    email: support@sync.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.sync.com/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or signup endpoints
  schemas:
...
    Post:
      type: object
      properties:
        postId:
          type: string
        communityId:
          type: string
        authorId:
          type: string
        type:
          type: string
          enum: [text, media, link]
        title:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [published, draft, removed]
        synergy:
          type: integer
        viewCount:
          type: integer
        commentCount:
          type: integer
        shareCount:
          type: integer
        saveCount:
          type: integer
        isNSFW:
          type: boolean
        isSpoiler:
          type: boolean
        isStickied:
          type: boolean
        isLocked:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Comment:
      type: object
      properties:
        commentId:
          type: string
        postId:
          type: string
        authorId:
          type: string
        parentId:
          type: string
        content:
          type: string
        synergy:
          type: integer
        replyCount:
          type: integer
        level:
          type: integer
        path:
          type: string
        isEdited:
          type: boolean
        isPinned:
          type: boolean
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /post/create:
    post:
      tags:
        - Post
      summary: Create a new post
      description: |
        Create a new post in a community. Supports text, media, and link posts.

        **Important:**
        - Max 10 media files can be uploaded
        - Content-Type must be multipart/form-data when uploading files

        **Middleware:** AuthProvider, FileUpload(media - max 10 files)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - title
                - content
                - communityId
                - type
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 300
                  example: "My First Post"
                content:
                  type: string
                  minLength: 1
                  maxLength: 10000
                  example: "This is the content of my post"
                communityId:
                  type: string
                  example: "comm-123"
                type:
                  type: string
                  enum: [TEXT, IMAGE, VIDEO, LINK]
                  example: "TEXT"
                tags:
                  type: array
                  items:
                    type: string
                isNSFW:
                  type: boolean
                  default: false
                isSpoiler:
                  type: boolean
                  default: false
                media:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 10
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Post created successfully"
                  data:
                    $ref: '#/components/schemas/Post'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
  /post/get/{postId}:
    get:
      tags:
        - Post
      summary: Get post by ID
      description: Retrieve a post by its ID
      security:
        - BearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Post retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Post'
        '404':
          description: Post not found
  /post/{postId}:
    put:
      tags:
        - Post
      summary: Edit a post
      description: Edit an existing post (only by author)
      security:
        - BearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 300
                content:
                  type: string
                  maxLength: 10000
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Post updated successfully
        '403':
          description: Not authorized to edit this post
        '404':
          description: Post not found

    delete:
      tags:
        - Post
      summary: Delete a post
      description: Delete a post (only by author or moderator)
      security:
        - BearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Post deleted successfully
        '403':
          description: Not authorized to delete this post
        '404':
          description: Post not found
  /post/like/{postId}:
    post:
      tags:
        - Post
      summary: Like a post
      description: Add a like to a post
      security:
        - BearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Post liked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Post liked"
                  data:
                    type: object
                    properties:
                      synergy:
                        type: integer
  /post/dislike/{postId}:
    post:
      tags:
        - Post
      summary: Dislike a post
      description: Add a dislike to a post
      security:
        - BearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Post disliked successfully
  /post/save/{postId}:
    post:
      tags:
        - Post
      summary: Save a post
      description: Save a post to user's saved posts
      security:
        - BearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Post saved successfully
  /post/share/{postId}:
    post:
      tags:
        - Post
      summary: Share a post
      description: Increment share count for a post
      security:
        - BearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Post shared successfully
  /post/feed:
    get:
      tags:
        - Post
      summary: Get user feed posts
      description: Get personalized feed of posts from followed users and joined communities
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Feed posts retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
  /post/trending:
    get:
      tags:
        - Post
      summary: Get trending posts
      description: Get currently trending posts
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Trending posts retrieved
  /post/popular:
    get:
      tags:
        - Post
      summary: Get popular posts
      description: Get most popular posts
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Popular posts retrieved
  /post/saved:
    get:
      tags:
        - Post
      summary: Get saved posts
      description: Get user's saved posts
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Saved posts retrieved
  /post/get/community/{communityId}:
    get:
      tags:
        - Post
      summary: Get community posts
      description: Get all posts from a specific community
      security:
        - BearerAuth: []
      parameters:
        - name: communityId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Community posts retrieved

  # ==================== COMMENT ENDPOINTS ====================
  /comment/post/create:
    post:
      tags:
        - Comment
      summary: Create a comment on a post
      description: |
        Add a comment to a post

        **Middleware:** AuthProvider, LocationProvider
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - postId
                - content
              properties:
                postId:
                  type: string
                content:
                  type: string
                  minLength: 1
                  maxLength: 5000
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Comment'
  /comment/post/{commentId}:
    put:
      tags:
        - Comment
      summary: Edit a comment
      description: Edit an existing comment (only by author)
      security:
        - BearerAuth: []
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  maxLength: 5000
      responses:
        '200':
          description: Comment updated successfully

    delete:
      tags:
        - Comment
      summary: Delete a comment
      description: Delete a comment (only by author or moderator)
      security:
        - BearerAuth: []
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Comment deleted successfully
  /comment/post/{postId}:
    get:
      tags:
        - Comment
      summary: Get comments for a post
      description: Retrieve all top-level comments for a post
      security:
        - BearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Comments retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Comment'
...